IMMICH/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monorepo-framework"
version = "1.0.0"
description = "Multi-project Python monorepo framework"
authors = [{name = "Your Name", email = "your.email@example.com"}]
license = {text = "MIT"}
requires-python = ">=3.8"

# Common dependencies used across projects
dependencies = [
    "invoke>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0.0",  # <-- add this line
    "requests>=2.0.0",
    "piexif",
    "Pillow",
]

# Optional dependency groups
[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "Pillow>=10.0.0",  # For image processing and test image generation
]
build = [
    "setuptools>=65.0.0",
    "wheel>=0.38.0",
    "build>=0.10.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
pythonpath = ["scripts", "src","common/src"]
markers = [
    "integration: mark a test as an integration test"
]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'